Fix some bugs to do with 100% scrobble point - we assume a track within 1 second of completion got to 100% - only allow the user to set scrobble point to 5% accuracy - don't do the thing with m_trackToScrobble (looks like old code that once made sense, but not anymore) - definitely don't scrobble the same track twice (extra check) - move scrobble point setting (and other related) to the user settings class

+++ app/client/Settings/ScrobbleSettingsWidget.h
@@ -41,7 +41,7 @@ private slots:
 
 private:
     Ui::ScrobbleSettingsWidget* ui;
-    int m_initialScrobblePercentage;
+    double m_initialScrobblePercentage;
 };
 
 #endif
